DRF_Study_Tutorial/mini_crud_project/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15.0

